+++ app/Http/Controllers/ThuctapsinhController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Http\Requests\StudentRequest;
+use Illuminate\Support\Facades\Gate;
 use App\Models\User;
 use App\Models\dot_thuctap;
 use App\Models\chucvu;
@@ -20,6 +21,7 @@ class ThuctapsinhController extends Controller
     {
         $title = "Danh sách thực tập sinh";
         $data = User::all();
+        // $data = DB::table('users')->join('dot_thuctap','users.id_dot','=','dot_thuctap.id_dot')->select();
         $get_dotthuctap = dot_thuctap::all();
         $get_chucvu = chucvu::all();
         return view('thuctapsinh.list', compact('title','data','get_dotthuctap','get_chucvu'));
@@ -32,10 +34,13 @@ class ThuctapsinhController extends Controller
      */
     public function create()
     {
+        if (Gate::allows('get-quantrivien')) {
         $title = "Thêm thực tập sinh";
         $get_dotthuctap = dot_thuctap::all();
         $get_chucvu = chucvu::all();
-      
+        } else {
+            return abort(403);
+        }
         return view('thuctapsinh.add', compact('title','get_dotthuctap','get_chucvu'));
     }
 
@@ -89,11 +94,14 @@ class ThuctapsinhController extends Controller
      */
     public function edit($id_sv)
     {
+        if (Gate::allows('get-quantrivien')) {
         $title = "Cập nhật thực tập sinh";
         $get_dotthuctap = dot_thuctap::all();
         $get_chucvu = chucvu::all();
         $t= User::find($id_sv);
-        
+        } else {
+            return abort(403);
+        }
         return view('thuctapsinh.edit',compact('t','title','get_chucvu','get_dotthuctap'));
     }
 
@@ -136,8 +144,12 @@ class ThuctapsinhController extends Controller
      */
     public function destroy($id_sv)
     {
+        if (Gate::allows('get-quantrivien')) {
         $t= User::find($id_sv);
         $t->delete();
+        } else {
+            return abort(403);
+        }
         return redirect()->back()->with(['success' => 'Xóa thành công !']);
     }
 }
